requirements.txt
numpy
opencv-python==4.5.5.64
Pillow
protobuf==3.19.0
PyYAML==6.0
tensorboardX==2.5
tqdm==4.64.0
# cpu
torch==1.11.0
torchaudio==0.11.0
torchvision==0.12.0
# #gpu
# torch==1.11.0+cu113
# torchaudio==0.11.0+cu113
# torchvision==0.12.0+cu113
